LunchLunch/src/com/lunchlunch/model/person/Person.java
package com.lunchlunch.model.person;

public class Person implements PersonInterface {

	private String id;
	private String firstName;
	private String lastName;
	private String email;

	public Person(String id, String firstName, String lastName, String email) {
		this.id = id;
		this.firstName = firstName;
		this.lastName = lastName;
		this.email = email;
	}

	@Override
	public String getFirstName() {
		return firstName;
	}

	@Override
	public String getLastName() {
		return lastName;
	}

	@Override
	public String getEmail() {
		return email;
	}

	@Override
	public String getId() {
		return id;
	}

}
